pyproject.toml
[tool.poetry]
name = "imgapp"
version = "0.1.0"
description = "Tweag poetry2nix example"
authors = ["Peter Dragos <peter@mlabs.city>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
Pillow = "^9.2.0"
Flask = "^2.2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
